Dev kmr

+++ workshop.pl
@@ -1141,16 +1141,21 @@ foreach my $fs (@virtual_fs) {
 
 if (-e $container_mount_point . "/etc/resolv.conf") {
     logger('info', "Backing up the temporary container's /etc/resolv.conf...\n");
+    my $command_output_log = "";
     ($command, $command_output, $rc) = run_command("/bin/cp --verbose --force " . $container_mount_point . "/etc/resolv.conf " . $container_mount_point . "/etc/resolv.conf.workshop");
-    ($command, $command_output, $rc) = run_command("/bin/rm --verbose --force " . $container_mount_point . "/etc/resolv.conf");
+    $command_output_log .= sprintf($command_logger_fmt, $command, $rc, $command_output);
+    if ($rc == 0) {
+        ($command, $command_output, $rc) = run_command("/bin/rm --verbose --force " . $container_mount_point . "/etc/resolv.conf");
+        $command_output_log .= sprintf($command_logger_fmt, $command, $rc, $command_output);
+    }
     if ($rc != 0) {
         logger('info', "failed\n", 1);
-        command_logger('error', $command, $rc, $command_output);
+        logger('error', $command_output_log);
         logger('error', "Failed to backup the temporary container's /etc/resolv.conf!\n");
         exit(get_exit_code('resolve.conf_backup'));
     }
     logger('info', "succeeded\n", 1);
-    command_logger('verbose', $command, $rc, $command_output);
+    logger('verbose', $command_output_log);
 }
 
 logger('info', "Temporarily copying the host's /etc/resolv.conf to the temporary container...\n");
@@ -1488,7 +1493,7 @@ if (opendir(NORMAL_ROOT, "/")) {
                             exit(get_exit_code('cpanm_install_failed'));
                         }
                     }
-                    logger('info', "succeeded\n", 2);
+                    logger('info', "succeeded\n", 4);
                     logger('verbose', $cpan_install_log);
                   } elsif ($req->{'type'} eq 'node') {
                       logger('info', "installing package via npm install...\n", 2);
@@ -1506,7 +1511,7 @@ if (opendir(NORMAL_ROOT, "/")) {
                               exit(get_exit_code('npm_install_failed'));
                           }
                       }
-                      logger('info', "succeeded\n", 2);
+                      logger('info', "succeeded\n", 4);
                       logger('verbose', $npm_install_log);
                   } elsif ($req->{'type'} eq 'python3') {
                       logger('info', "installing package via python3 pip...\n", 2);
@@ -1524,7 +1529,7 @@ if (opendir(NORMAL_ROOT, "/")) {
                               exit(get_exit_code('python3_install_failed'));
                           }
                       }
-                      logger('info', "succeeded\n", 2);
+                      logger('info', "succeeded\n", 4);
                       logger('verbose', $python3_install_log);
                 }
             }
